# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentación Colegio Catolico Api",
    "version": "1.0.0",
    "description": "Documentación de la API de gestor de pagos de Colegio Catolico"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
...
    "/leveling/listarLeveling/{codigo}": {
      "get": {
        "summary": "Obtiene un curso de nivelación por código.",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "codigo",
            "required": true,
            "description": "Código del curso de nivelación.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del curso de nivelación."
          },
          "404": {
            "description": "Curso de nivelación no encontrado."
          }
        }
      }
    },
    "/leveling/listarEstudiante/{codigo}": {
      "get": {
        "summary": "Lista datos del estudiante por código.",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "codigo",
            "required": true,
            "description": "Código del estudiante.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del estudiante.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/leveling/listarLevelingId/{id}": {
      "get": {
        "summary": "Obtiene un curso de nivelación por ID.",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del curso de nivelación.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del curso de nivelación."
          },
          "404": {
            "description": "Curso de nivelación no encontrado."
          }
        }
      }
    },
    "/leveling/crearLeveling": {
      "post": {
        "summary": "Crear un nuevo proceso de nivelación",
        "tags": [
          "Leveling"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modalidadCurso": {
                    "type": "string",
                    "description": "Modalidad del curso de nivelación"
                  },
                  "asignatura": {
                    "type": "string",
                    "description": "Asignatura en la que se realizará la nivelación"
                  },
                  "isActive": {
                    "type": "boolean",
                    "description": "Estado activo del proceso de nivelación (true o false)"
                  },
                  "estadoAprobado": {
                    "type": "string",
                    "description": "Estado de aprobación del proceso"
                  },
                  "grado": {
                    "type": "string",
                    "description": "Grado del estudiante"
                  },
                  "idEstudiante": {
                    "type": "string",
                    "description": "ID del estudiante que realiza la nivelación"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nivelación creada con éxito"
          },
          "400": {
            "description": "Error en los datos enviados"
          }
        }
      }
    },
    "/leveling/actualizarLeveling/{id}": {
      "put": {
        "summary": "Actualizar un proceso de nivelación existente",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID del proceso de nivelación a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modalidadCurso": {
                    "type": "string",
                    "description": "Modalidad del curso de nivelación"
                  },
                  "asignaturaisActive": {
                    "type": "boolean",
                    "description": "Estado activo de la asignatura en el proceso de nivelación"
                  },
                  "estadoAprobado": {
                    "type": "string",
                    "description": "Estado de aprobación del proceso"
                  },
                  "grado": {
                    "type": "string",
                    "description": "Grado del estudiante"
                  },
                  "idEstudiante": {
                    "type": "string",
                    "description": "ID del estudiante que realiza la nivelación"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nivelación actualizada con éxito"
          },
          "400": {
            "description": "Error en los datos enviados"
          }
        }
      }
    },
    "/leveling/deshabilitar/{id}": {
      "put": {
        "summary": "Deshabilita un curso de nivelación por ID.",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del curso de nivelación.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean",
                    "description": "Estado de activación del curso de nivelación."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Curso de nivelación deshabilitado."
          },
          "400": {
            "description": "Error en los datos enviados."
          }
        }
      }
    },
    "/leveling/cambiarEstado/{id}": {
      "put": {
        "summary": "Cambia el estado de aprobación de un curso de nivelación por ID.",
        "tags": [
          "Leveling"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID del curso de nivelación.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estadoAprobado": {
                    "type": "string",
                    "description": "Estado de aprobación del curso de nivelación."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado del curso de nivelación cambiado."
          },
          "400": {
            "description": "Error en los datos enviados."
          }
        }
      }
    },
...
  },
...
}